requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
av
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
cors
distro
docstring_parser
edge-tts
eval_type_backport
fastapi
filelock
Flask
flask-cors
flatbuffers
frozenlist
fsspec
future
gevent
google-api-core
google-auth
google-cloud-speech
google-cloud-texttospeech
google-genai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
itsdangerous
Jinja2
jiter
livekit
livekit-agents
livekit-api
livekit-blingfire
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-protocol
MarkupSafe
mpmath
multidict
nest-asyncio
networkx
numpy
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
packaging
pillow
prometheus_client
propcache
proto-plus
protobuf
psutil
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
PyJWT
pyreadline3
PySocks
python-dotenv
PyYAML
regex
requests
requests-file
rsa
safetensors
setuptools
sniffio
sounddevice
starlette
sympy
tabulate
tenacity
tldextract
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
types-protobuf
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
Werkzeug
yarl
zipp
zope.event
zope.interface